Fixes and add frontend to docker-compose

+++ usersapi/src/main/java/org/uniara/usersapi/controller/UserController.java
@@ -6,6 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.uniara.usersapi.constant.Constant;
 import org.uniara.usersapi.model.User;
+import org.uniara.usersapi.security.JwtTokenProvider;
 import org.uniara.usersapi.service.UserService;
 
 import java.util.List;
@@ -16,28 +17,56 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private JwtTokenProvider jwtTokenProvider;
+
     @GetMapping(Constant.API_USERS)
-    public ResponseEntity<List<User>> findAll(/*@RequestHeader("Authorization") String token*/) {
+    public ResponseEntity<List<User>> findAll(@RequestHeader("Authorization") String token) {
+
+        if (!jwtTokenProvider.validateToken(token.replace("Bearer ", ""))) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         return ResponseEntity.ok(userService.findAll());
     }
 
     @GetMapping(Constant.API_USERS + "/{id}")
-    public ResponseEntity<Optional<User>> findById(/*@RequestHeader("Authorization") String token,*/ @PathVariable("id") Long id) {
+    public ResponseEntity<Optional<User>> findById(@RequestHeader("Authorization") String token, @PathVariable("id") Long id) {
+
+        if (!jwtTokenProvider.validateToken(token.replace("Bearer ", ""))) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         return ResponseEntity.ok(userService.findById(id));
     }
 
     @PostMapping(Constant.API_USERS)
-    public ResponseEntity<User> save(/*@RequestHeader("Authorization") String token,*/ @RequestBody User user) {
+    public ResponseEntity<User> save(@RequestHeader("Authorization") String token, @RequestBody User user) {
+
+        if (!jwtTokenProvider.validateToken(token.replace("Bearer ", ""))) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         return ResponseEntity.status(HttpStatus.CREATED).body(userService.save(user));
     }
 
     @PutMapping(Constant.API_USERS)
-    public ResponseEntity<User> update(/*@RequestHeader("Authorization") String token,*/ @RequestBody User user) {
+    public ResponseEntity<User> update(@RequestHeader("Authorization") String token, @RequestBody User user) {
+
+        if (!jwtTokenProvider.validateToken(token.replace("Bearer ", ""))) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         return ResponseEntity.ok(userService.save(user));
     }
 
-    @DeleteMapping(Constant.API_USERS + "{id}")
-    public ResponseEntity<User> delete(/*@RequestHeader("Authorization") String token,*/ @PathVariable("id") Long id) {
+    @DeleteMapping(Constant.API_USERS + "/{id}")
+    public ResponseEntity<User> delete(@RequestHeader("Authorization") String token, @PathVariable("id") Long id) {
+
+        if (!jwtTokenProvider.validateToken(token.replace("Bearer ", ""))) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         userService.deleteById(id);
         return ResponseEntity.noContent().build();
     }
